==> 1_loading_before_after_eps_days.ipynb <==
# Project Goal:

PREDICT percent change in stock price after next earnings event for all company

One record per earning event date per company.

## Document Purpose:
Get the day before and day after for each stock's earning events (before_eps_day, after_eps_day)

# Packages Imports

!pip install pandas_market_calendars

# Loading Data

## Filter Trading Dates for tickers

Based on tickers in ticker csv, only keep earnings dates records for those stocks

## Imports
import pandas as pd
import pandas_market_calendars as mcal
import numpy as np
from datetime import datetime
import seaborn as sns
import matplotlib.pyplot as plt
import plotly.express as px

import yfinance as yf

# list of stock tickers we are interested in measuring the percent change in (provided):
tickers_df = pd.read_csv("/users/brigitteasullivan/Documents/0. Data Science/Notebooks/Riipen Project/eps-stock-market-ml-project/data/all_tickers-with-options.csv", 
                         header = 0, names = ['Ticker'])
tickers_df

# cleaning: trim tab
tickers_df["Ticker"] = tickers_df['Ticker'].str.strip()
tickers_df

## Earnings dates

Objective is to get a percentage change in stock price between two dates, to then calculate the percentage change in stock price after an earnings announcement. Companies announce earnings either "premarket", where earnings are announced before markets open, or "postmarket" after markets close for that day. 

**Decision:** If company announced is premarket - earnings date to watch is same day and the previous open market day. 
if company announced is postmarket - earnings date to watch is next trading date and the day anounced.

From documentation: https://www.earningscalendar.net/documentation

*   `date` ISO date of this earnings event.
*   `ticker` Ticker as traded on the exchange.
*   `url` URL to the press release that confirmed this event.
*   `title` Title of the press release that confirmed this event.
*   `security\_name` Name of the security trading under this symbol/ticker.
*   `exchange` The exchange where the security is trading. One of AMEX, NASDAQ, NYSE.
*   `when` When are earnings slated to be released that day. Returned only if known, otherwise the string "null".
*   `pub\_date` Publication date of the press release that confirmed this event



# load provided data set for earnings events:
earnings_df = pd.read_csv('/users/brigitteasullivan/Documents/0. Data Science/Notebooks/Riipen Project/eps-stock-market-ml-project/data/confirmed_earnings.csv')
earnings_df

earnings_df = earnings_df.loc[:,['date', 'exchange', 'symbol', 'when', 'security_name']].copy()

earnings_df

**Decision:** If no value for when company announced, assume the announcement is premarket. 

# cleaning - remove nans:
# fill missing NA's in "when" with "premarket"

earnings_df.loc[:,'when'].fillna(value='premarket', inplace=True)
earnings_df

# Convert the 'date' column to datetime
earnings_df['date'] = pd.to_datetime(earnings_df['date'])

# QA - check date data type
earnings_df.dtypes

# remove data for companies not in ticker list
earnings_df = earnings_df[earnings_df['symbol'].isin(tickers_df['Ticker'])].copy()
earnings_df

# QA: check how many tickers remain
ticker_unique_num = tickers_df['Ticker'].nunique()
print('unique ticker symbols in ticker df:', ticker_unique_num)

symbol_unique_num = earnings_df['symbol'].nunique()
print('unique symbols in earnings df:', symbol_unique_num)


# in both ticker and earnings df
combined_df = pd.concat([earnings_df['symbol'], tickers_df['Ticker']])
unique_ticker = set(tickers_df['Ticker'].unique())
unique_earnings = set(earnings_df['symbol'].unique())

intersection = list(unique_ticker.intersection(unique_earnings))
length_intersection = len(intersection)
print('unique symbols in earnings df AND in ticker:', length_intersection)

earnings_df['exchange'].value_counts()

# check percent missing values for each column
percent_missing = earnings_df.isnull().sum() * 100 / len(earnings_df)
missing_value_df = pd.DataFrame({'column_name': earnings_df.columns,
                                 'percent_missing': percent_missing})

missing_value_df

## cleaning - filter out OTC

earnings_df = earnings_df.loc[~earnings_df['exchange'].isin(["OTC"]),:].copy()
# earnings_df

## QA - check that OTC not in Dataframe anymore
earnings_df['exchange'].value_counts(dropna = False)

earnings_df

# Before and After Earnings Dates
Find the next trading date based on earnings date release and stock market the company is trading on. 


Note:
AMEX stock exchange is also known as the NYSE American and follows the same open schedule as the NYSE, Will use the NYSE calendar for all AMEX values.

# dictionary of stock exchanges and codes 

stock_ex_codes = {
    'NYSE': 'XNYS',
    'New York Stock Exchange (NYSE)': 'XNYS',
    'NASDAQ': 'NASDAQ',
    'London Stock Exchange (LSE)': 'XLON',
    'Tokyo Stock Exchange (TSE)': 'XTKS',
    'Hong Kong Stock Exchange (HKEX)': 'XHKG',
    'Toronto Stock Exchange (TSX)': 'XTSE',
    'Shanghai Stock Exchange (SSE)': 'XSHG',
    'Shenzhen Stock Exchange (SZSE)': 'XSHE',
    'Bombay Stock Exchange (BSE)': 'XBOM',
    'National Stock Exchange of India (NSE)': 'XNSE',
    'AMEX': 'XNYS' # See note
}


# Create a copy of the 'exchange' column
exchange_column_copy = earnings_df['exchange'].copy()

# Add a new column 'exchange_code' based on the copied 'exchange' column
earnings_df.loc[:,'exchange_code'] = exchange_column_copy.map(stock_ex_codes)

earnings_df

earnings_df['exchange'].value_counts(dropna=False)

#drop rows that didn't have a exchange that was found:
earnings_df = earnings_df.dropna(inplace = False, ignore_index = True, subset=['exchange', 'security_name', 'exchange_code']).copy()
earnings_df

earnings_df['exchange'].value_counts(dropna=False)

# store calendars in seperate variables to increase speed - function + apply was 15 mins
nyse_cal = mcal.get_calendar('XNYS')
nasdaq_cal = mcal.get_calendar('NASDAQ')


#### Trading Day After Earnings Release

def get_next_trading_day(row):
    ''' Returns the next trading day that a specific exchange is open based on a input date, considering holidays / weekends.
    If no value for exchange return np.nan. 
    will be applied to a df.
    '''
    next_trading_day2 = np.nan
    try:
        if row['exchange_code'] == 'XNYS':
            next_trading_day2 = nyse_cal.valid_days(start_date=row['date'], end_date = row['date'] +pd.Timedelta(days=10))[1]
        elif row['exchange_code'] == 'NASDAQ':
            next_trading_day2 = nasdaq_cal.valid_days(start_date=row['date'], end_date = row['date'] +pd.Timedelta(days=10))[1]
        # exchange =  mcal.get_calendar(row['exchange'])
        # next_trading_day2 = exchange.valid_days(start_date=row['date'], end_date = row['date'] +pd.Timedelta(days=10))[1]
        return next_trading_day2
    except Exception as e:
        print(f'Error:{e}')
        return np.nan

#show how long this cell took to run.
import timeit

start_time = timeit.default_timer()

#apply function
earnings_df.loc[:,'next_trading_day'] = earnings_df.apply(get_next_trading_day, axis = 1)

#print lenght of time to run function on entire DF
elapsed = timeit.default_timer() - start_time
print(f"Time taken: {elapsed} seconds")

earnings_df

# QA: check to make sure all records in df have a next trading day
earnings_df[earnings_df['next_trading_day'].isnull()]

#drop rows that didn't have a trading day that was found (there should be none)
earnings_df = earnings_df.dropna(inplace = False, ignore_index = True).copy()
earnings_df

# store as datetime types
earnings_df['next_trading_day'] = pd.to_datetime(earnings_df['next_trading_day'])
earnings_df['date'] = pd.to_datetime(earnings_df['date'])
earnings_df.dtypes

### Trading Day Before Earnings release

def get_prior_trading_day(row):
    ''' Returns the trading day PRIOR that a specific exchange is open based on a input date, considering holidays / weekends.
    If no value for exchange return np.nan. 
    will be applied to a df.
    start date is 10 days before known trading date, then take second last item from list (last item is the end date) 
    10 days to make sure there's enough range for winter holidays
    '''
    try:
        if row['exchange_code'] == 'XNYS':
            prior_trading_day = nyse_cal.valid_days(start_date = row['date'] -pd.Timedelta(days=10), end_date=row['date'])[-2]
        elif row['exchange_code'] == 'NASDAQ':
            prior_trading_day = nasdaq_cal.valid_days(start_date = row['date'] -pd.Timedelta(days=10), end_date=row['date'])[-2]
        return prior_trading_day
        # exchange =  mcal.get_calendar(row['exchange_codes'])
        # prior_trading_day = exchange.valid_days(start_date = row['date'] -pd.Timedelta(days=10), end_date=row['date'])[-2]
        # return prior_trading_day
    except Exception as e:
        # print(f'Error:{e}')
        return np.nan

#show how long this cell took to run.
import timeit

start_time = timeit.default_timer()

earnings_df['prior_trading_day'] = earnings_df.apply(get_prior_trading_day, axis = 1)

elapsed = (timeit.default_timer() - start_time) / 60
print(f"Time taken: {elapsed} minutes")

earnings_df

### Next trading day or day of 

# QA Confirm no NaN in When column
earnings_df['when'].value_counts(dropna=False)

## ASSUMPTION - IF NO specification on when, assuming pre market and using same day as release as the eps day (per direction from manager)

def eps_day_after(row):
    eps_day_value = np.nan
    if row['when'] == 'premarket':
        eps_day_value = row['date']
    elif row['when'] == 'postmarket':
         eps_day_value = row['next_trading_day']
    elif pd.isna(row['when']):
        eps_day_value = row['date']
    return eps_day_value

## Based on when the EPS is announced, determine the day we are interested in for movement
earnings_df['after_eps_date'] = earnings_df.apply(eps_day_after, axis = 1)

earnings_df

# Define a function to conditionally replace values
def eps_day_before(row):
    if row['when'] == 'postmarket':
        return row['date']
    elif row['when'] == 'premarket':
        return row['prior_trading_day']
    else:
        return row['next_trading_day']

# Apply the function to create a new column 'D' with the replaced values
earnings_df['before_eps_date'] = earnings_df.apply(eps_day_before, axis=1)


earnings_df['after_eps_date'] = pd.to_datetime(earnings_df['after_eps_date'], utc=True).dt.strftime('%Y-%m-%d')
earnings_df['before_eps_date'] = pd.to_datetime(earnings_df['before_eps_date'], utc=True).dt.strftime('%Y-%m-%d')

earnings_df.head(15)

# tidy up final data set, drop unecessary columns, reorder
earnings_df  = earnings_df.drop(columns = ['next_trading_day', 'prior_trading_day', 'exchange'], inplace=False).copy()

earnings_df = earnings_df.reindex(columns = ['date', 'symbol', 'when', 'security_name', 'exchange_codes', 'before_eps_date', 'after_eps_date'])
earnings_df

# Export

#export to CSV as EPS DAYS
earnings_df.to_csv('/users/brigitteasullivan/Documents/0. Data Science/Notebooks/Riipen Project/eps-stock-market-ml-project/data/eps_days.csv', index = False)
