==> 2_Data_Loading_Part2.ipynb <==
# Package imports

import pandas as pd
import pandas_market_calendars as mcal
import numpy as np
from datetime import datetime
import seaborn as sns
import matplotlib.pyplot as plt
import plotly.express as px

import yfinance as yf

# Stock history

events_df = pd.read_csv('data/eps_days.csv')
events_df.head(10)

print("DF length:", len(events_df))

events_df.dtypes

events_df['date'] = pd.to_datetime(events_df['date'])
events_df['after_eps_date'] = pd.to_datetime(events_df['after_eps_date'])
events_df['before_eps_date'] = pd.to_datetime(events_df['before_eps_date'])

unique_tickers = events_df['symbol'].unique().tolist()
print("number of unique tickers:", len(unique_tickers))
print("type:", type(unique_tickers))

# get smallest before date: 
min_date = events_df['before_eps_date'].min()

print(f'earilest before date : {min_date}')
# get largest after date: 
max_date = events_df['after_eps_date'].max()

print(f'latest after date : {max_date}')

# check tickers 
multi_data = yf.download(unique_tickers, start = '2022-12-30', end = '2023-12-30', interval='1d')
multi_data

multi_data.to_csv('data/stock_price_history.csv', index = False)

# multi_data = pd.read_csv('data/stock_price_history.csv', low_memory = False, header='infer'
#                     )

# test = multi_data.reset_index(drop=False).copy()


multi_data.T.index

multi_data.columns = multi_data.columns.map('_'.join)

multi_data = multi_data.T.reset_index().copy()
multi_data

# Split the 'Column' into two columns
multi_data[['Status', 'Ticker']] = multi_data['index'].str.split('_', expand=True)

multi_data = multi_data.drop(columns = 'index')
multi_data

multi_data = multi_data.melt(id_vars=['Status', 'Ticker'])


multi_data = multi_data.rename_axis('Index')
multi_data

multi_data

clean_df = multi_data.pivot(index = ['Ticker', 'Date'], columns='Status', values='value').copy()
clean_df = clean_df.reset_index()

clean_df.to_csv('data/stock_price_history_clean.csv', index = False)

prices_df = pd.read_csv('data/stock_price_history_clean.csv')
prices_df

prices_df.sort_values(by = ['Ticker', 'Date'], ascending = True)

stock_status = []
for num in range(len(cols)):
        stock_status.append(cols[num][0])
    

stock_status_list = list(set(stock_status))
stock_status_list

stock_result = pd.DataFrame()
def stock_history(row):
    ''' get stock prices for the days between before and after for each stock'''
    ticker = row['symbol']
    # print(ticker)
    data = yf.Ticker(ticker)
    result = data.history(start=row['before_eps_date'], end = row['after_eps_date']+pd.Timedelta(days=1))
    result['symbol'] = ticker
    result.reset_index(drop=False, inplace=True)
    result.set_index('symbol', inplace =True)
    # result['Date'] = pd.to_datetime(result['Date'])
    result['Date'] = pd.to_datetime(result['Date']).dt.strftime('%Y-%m-%d')
    return result
    
    

# QA testing one row
# pct_test = stock_history(events_df.iloc[1,:])

# pct_test

# full DF
prices = events_df.apply(stock_history, axis=1)
prices

pct_test.dtypes


pct_test = pct_test[['Open', 'High','Low','Close']].pct_change(axis=0).dropna()



pct_test

pct_test = pct_test[['Open', 'High','Low','Close']].pct_change(axis=0).dropna()
    
    

## Bin

ticker_list = events_df['symbol'].tolist()
ticker_list = set([ticker.lower() for ticker in ticker_list])
print(len(ticker_list))
ticker_list

print(events_df.loc[0,'after_eps_date'])
print(events_df.loc[0,'after_eps_date']+pd.Timedelta(days=1))

events_df['after_eps_date'] = np.nan
events_df['before_eps_date'] = np.nan

events_df

### next trading day is after eps date for all values where WHEN = Postmarket

# Define a function to conditionally replace values
def get_after_date(row):
    if row['when'] == 'postmarket':
        return row['next_trading_day']
    elif row['when'] == 'premarket':
        return row['date']
    else:
        return row['next_trading_day']

# Apply the function to create a new column 'D' with the replaced values
events_df['after_eps_date'] = events_df.apply(get_after_date, axis=1)



events_df

# Define a function to conditionally replace values
def get_before_date(row):
    if row['when'] == 'postmarket':
        return row['date']
    elif row['when'] == 'premarket':
        return row['prior_trading_day']
    else:
        return row['next_trading_day']

# Apply the function to create a new column 'D' with the replaced values
events_df['before_eps_date'] = events_df.apply(get_before_date, axis=1)




events_df.dtypes

events_df['after_eps_date'] = pd.to_datetime(events_df['after_eps_date'])
events_df['before_eps_date'] = pd.to_datetime(events_df['before_eps_date'])

events_df.head(15)

events_df.dtypes
