==> 3_loading_earnings_amounts.ipynb <==
# Document Purpose
* get quarterly EPS values for each ticker on each earnings events dates

# Imports

import pandas as pd
import yfinance as yf
import numpy as np

events_df = pd.read_csv("data/exports/stock_price_pct_change_history.csv")

# view df
events_df

# drop unecessary columns
events_df = events_df[['date', 'symbol','security_name', 
       'before_eps_date', 'after_eps_date', 'Adj Close',
       'Close',  'Adj Close_pct_change',
       'Close_pct_change']].copy()

events_df

result_eps = pd.DataFrame()
def get_basic_eps(row):  
    result = pd.DataFrame()
    result = pd.DataFrame(columns = ['Earnings Date', 'ticker', "EPS Estimate", "Reported EPS", "Surprise(%)"])
    print(row)
    return result
    try:
        ticker_data = yf.Ticker(row)
        eps_data  = ticker_data.get_earnings_dates(limit=12)
        # income_stmt = ticker_data.income_stmt.T["Basic EPS"]
        result = pd.DataFrame(eps_data).reset_index()
        result['ticker'] = row
        result = result[['Earnings Date', 'ticker', "EPS Estimate", "Reported EPS", "Surprise(%)"]].copy()
        result.dropna(inplace = True)
        pd.concat(result_eps, result)
        return result
        
    except Exception as e:
        print(f'Error:{e}')
        return np.nan

# Check Tickers 

confirm that yfinance has ticker earnings data available

unique_tickers = pd.DataFrame(events_df['symbol'].unique(), columns = ['ticker'])

unique_tickers

# QA - test on subset 
# unique_tickers_test = unique_tickers.iloc[:20, :].copy()
# unique_tickers_test

def check_ticker(row):
    """for each ticker, get earnings data if no earnings data found for the stock, return the error and the symbol"""
    errors_list = [] # to keep track of what happened to each stock
    result = pd.DataFrame()
    try:
        ticker_data = yf.Ticker(row)
        eps_data  = ticker_data.get_earnings_dates(limit=12).reset_index()
        eps_data['Symbol'] = row # add a column with the symbol into the DF (to concat later)
        # summary_list.append({'Symbol': row, 'Errors': 'Earnings Data found'})
        result = eps_data
    except Exception as e:
        result = {'Symbol': row, 'Errors': f"{e}"}
    res_dict = [result, errors_list]
    return result

Warning: the below cell can take a long time to run.

import timeit

start_time = timeit.default_timer()

# get earnings data for each ticker: 
tuple = unique_tickers['ticker'].apply(check_ticker)

## Time:
elapsed_seconds = timeit.default_timer() - start_time
elapsed_minutes = elapsed_seconds / 60
print(f"Time taken: {elapsed_seconds} seconds")
print(f"Time taken: {elapsed_minutes} minutes")

print(type(tuple))
print(len(tuple))
print(tuple)

tuple

final_df = pd.DataFrame()
# where a dict is found, there is EPS data available. Store dict values in final_df
for index in range(len(tuple)):
    item = tuple[index]
    # print(item)
    result_df = []
    error_df_test = pd.DataFrame()
    if isinstance(item, pd.DataFrame):
        final_df = pd.concat([final_df,item])
    elif isinstance(item, dict):
        print('found dict:')
        print(item)
        # error_df_test = pd.DataFrame(item, index = [index])
        # errors_df = pd.concat([error_df_test,tuple[index]])
    else:
        print("Unknown type")
# print(error_df_test)
print(final_df)


final_df

len(final_df["Symbol"].unique())

# export earnings per share amounts 
final_df.to_csv("data/exports/eps_data.csv", index = False)

final_df
