==> 4_merge_eps_percent_change.ipynb <==
# Document Purpose:
* purpose of this notebook is to merge the stock price data on particular earnings announcement dates with the EPS data for those dates into one dataframe


# Imports

import pandas as pd
import numpy as np

# Cleaning

events_df = pd.read_csv("data/exports/stock_price_pct_change_history.csv")
# drop unecessary columns
events_df = events_df[['date', 'symbol','security_name', 'when',
       'before_eps_date', 'after_eps_date', 'Adj Close',
       'Close',  'Adj Close_pct_change',
       'Close_pct_change']].copy()

eps_data = pd.read_csv("data/exports/eps_data.csv")

eps_data

events_df[events_df["symbol"] == "SGH"]
# events_df.head()

eps_data[eps_data["Symbol"] == "SGH"]

#how many rows where reported EPS is NaN: ~= 7803 rows
eps_data[eps_data[[ "Reported EPS"]].isna().all(axis=1)]



# drop rows where reported EPS is NaN:
eps_data = eps_data.dropna(axis = 0, subset= ["Reported EPS"], ignore_index = True).copy()
eps_data

eps_data[eps_data[['EPS Estimate', "Reported EPS", "Surprise(%)"]].isna().any(axis=1)]


eps_data.dtypes

# Convert 'date' column to datetime type
eps_data["Earnings Date"] = pd.to_datetime(eps_data["Earnings Date"], utc=True).dt.strftime('%Y-%m-%d')
print(eps_data.dtypes)



events_df.dtypes

# Merge EPS Data, Stock Data Pct Changes

Merge based on date and stock

events_df.sort_values(by = ["symbol", 'date'])

eps_data.sort_values(by = ["Symbol", 'Earnings Date'])

#### Decision: 
for earnings events where there is no matching earnings values, these event dates will be dropped.
The ML model requires complete data. 

merged_df = events_df.merge(eps_data, left_on = ['date', 'symbol'], right_on = ["Earnings Date", "Symbol"], how = "inner")
merged_df

 ### QA:
 confirming merge performed as expected by doing a full outer merge first, then reviewing results for individual stocks. 

 Result = if no earnings date is associated with the "date" field, ok to drop row, 

merged_df[merged_df['Earnings Date'].isnull()]

stock = "OGI"
single_stock_df = merged_df[(merged_df['symbol'] == stock) | (merged_df['Symbol'] == stock)]
single_stock_df

merged_df.sort_values(by=['symbol', 'Symbol', 'date', "Earnings Date"], na_position = 'last').head(2000)

merged_df.columns

# drop duplicate columns
merged_df.drop(columns =[ 'Earnings Date','Symbol'], inplace = True)

# convert when to binary 
merged_df.loc[merged_df['when'] == 'premarket','when'] = 1
merged_df.loc[merged_df['when'] == 'postmarket','when'] = 0
merged_df = merged_df.rename(columns = {'when': 'premarket_pub'}).copy()
merged_df

merged_df

#### Surprise % QA

## QA: check how surprise % is calculated. It should be (reported minus estimate) divided by estimate
merged_df['suprise_calc'] = (merged_df['Reported EPS'] - merged_df['EPS Estimate']) / merged_df['EPS Estimate']
merged_df['suprise_calc'] = merged_df['suprise_calc'].round(4).copy()

merged_df

### Decision: 
There is slight variation between the surprise % provided and the manually calculated surprise percentage. This will be attributed to rounding and the original values provided by yfinance will be used. 

merged_df.drop(columns=['suprise_calc'], inplace=True)

merged_df

# export merged df with complete data to csv
merged_df.to_csv("data/exports/data_full.csv", index = False)
